перерисовка табов
