+++ src/components/film-container/film-container.jsx
@@ -6,17 +6,13 @@ import ApiService from "../../store/api-actions";
 import {resetFilm} from "../../store/action";
 
 import Film from '../film/film';
-import FilmOverview from "../film-overview/film-overview";
-import FilmDetails from "../film-details/film-details";
-import FilmReviews from "../film-reviews/film-reviews";
 import LoadingScreen from '../loading-screen/loading-screen';
 
 const apiService = new ApiService();
 
-const FilmContainer = ({path, filmId}) => {
+const FilmContainer = ({filmId}) => {
 
   const currentFilm = useSelector(({FILM}) => FILM.currentFilm);
-  const currentFilmComments = useSelector(({FILM}) => FILM.currentFilmComments);
 
   const filteredFilms = useSelector(({ALL_FILMS}) => ALL_FILMS.filteredFilms);
   const filmsToShow = useSelector(({ALL_FILMS}) => ALL_FILMS.filmsToShow);
@@ -24,7 +20,6 @@ const FilmContainer = ({path, filmId}) => {
   const {authorizationStatus} = useSelector(({USER}) => USER);
 
   const loadFilmsData = () => dispatch(apiService.fetchFilm(filmId));
-
 
   const dispatch = useDispatch();
 
@@ -57,17 +52,6 @@ const FilmContainer = ({path, filmId}) => {
     backgroundImg
   } = film;
 
-  // создать компонент отдельный
-  const movieOverview = <FilmOverview film={film}/>;
-
-  const movieDetails = <FilmDetails film={film} />;
-
-  const movieReviews = <FilmReviews
-    reviews={currentFilmComments}
-    id={id}
-  />;
-
-
   const exactFilms = filteredFilms
     .filter(({genre}) => genre === filmGenre)
     .slice(0, 4);
@@ -85,17 +69,12 @@ const FilmContainer = ({path, filmId}) => {
       authorizationStatus={authorizationStatus}
       loadFilmsData={loadFilmsData}
       filmsToShow={filmsToShow}
-      movieOverview={movieOverview}
-      movieReviews={movieReviews}
-      movieDetails={movieDetails}
       filmId={filmId}
-      path={path}
-      id={id} />
+    />
   );
 };
 
 FilmContainer.propTypes = {
-  path: PropTypes.string.isRequired,
   filmId: PropTypes.number.isRequired,
 };
